+++ src/main/kotlin/uk/tldcode/learning/projecteuler/h0/t0/Problem6.kt
@@ -3,9 +3,9 @@ package uk.tldcode.learning.projecteuler.h0.t0
 
 class Problem6 {
 
-    fun sumSquareDiff(max: Int): Int {
-        var sum = 0
-        var square = 0
+    fun sumSquareDiff(max: Int): Long {
+        var sum = 0L
+        var square = 0L
         for (i in 1..max) {
             sum += i * i
             square += i
