+++ src/main/kotlin/uk/tldcode/learning/projecteuler/h0/t0/Problem9.kt
@@ -3,7 +3,7 @@ package uk.tldcode.learning.projecteuler.h0.t0
 
 class Problem9 {
 
-    fun specialPythagoreanTriplet(target: Double): Int {
+    fun specialPythagoreanTriplet(target: Double): Long {
         var a = 0
         var b = 0
         var c = fun(): Double {
@@ -15,12 +15,12 @@ class Problem9 {
             while (b < target) {
                 b++
                 if(a+b+c() == target){
-                    return (a * b * c()).toInt()
+                    return (a * b * c()).toLong()
                 }
             }
 
         }
-        return (a * b * c()).toInt()
+        return (a * b * c()).toLong()
 
 
     }
